refactor(constants): Devide contants so that now each page has local constants
