+++ src/utils/constantsItems.js
@@ -1,0 +1,8 @@
+/* constants for order of items */
+export const RECENT = "recent";
+export const FAVORITE = "favorite";
+
+export const ORDER_MESSAGE = {
+  [FAVORITE]: "좋아요순",
+  [RECENT]: "최신순",
+};
